refactor: move all tasks to one class
