+++ build.gradle
@@ -17,7 +17,7 @@ repositories {
 }
 
 dependencies {
-    implementation 'io.littlehorse:littlehorse-client:0.12.1'
+    implementation 'io.littlehorse:littlehorse-client:0.12.4'
     implementation 'org.slf4j:slf4j-api:2.0.7'
     implementation 'org.apache.logging.log4j:log4j-slf4j2-impl:2.20.0'
 }
